backend/pyproject.toml
[tool.poetry]
name = "food-tracker-api"
version = "0.1.0"
description = "Food Tracker API with AI-powered food recognition"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-multipart = "^0.0.6"
pillow = "^10.1.0"
requests = "^2.31.0"
python-dotenv = "^1.0.0"
supabase = "^2.0.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
httpx = "^0.25.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
black = "^23.11.0"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

